% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocSource-Data.R
\docType{data}
\name{socSource}
\alias{socSource}
\title{Social Mention source Data}
\format{An list with five formats:Blogs, MicroBlogs, Bookmarks, Images or all}
\usage{
data(socSource)
}
\description{
Social Mention source Data
}
\examples{
data(socSource)
source <- socSource[1]
}
\references{

}
\keyword{datasets}

