% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceType-Data.R
\docType{data}
\name{sourceType}
\alias{sourceType}
\title{Social Mention source Data}
\format{An list with five formats:Blogs, MicroBlogs, Bookmarks, Images or all}
\usage{
data(sourceType)
}
\description{
Social Mention source Data
}
\examples{
data(sourceType)
source <- sourceType[1]
}
\references{

}
\keyword{datasets}

