% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeFrame-Data.R
\docType{data}
\name{timeFrame}
\alias{timeFrame}
\title{Social Mention time frame Data}
\format{An data frame with six time intervals: h, 12h, 24h, w, m, all}
\usage{
data(timeFrame)
}
\description{
Social Mention time frame Data
}
\examples{
data(timeFrame)
allTime <- timeFrame[,1]
}
\references{

}
\keyword{datasets}

