% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocialMentions.R
\name{socialMention}
\alias{socialMention}
\title{Social Mentions}
\usage{
socialMention(term, source = "all", time.frame = "all")
}
\arguments{
\item{term}{- term which we intend to get the data}

\item{source}{- list of source of a term for which we intended to get the data}

\item{time.frame}{- get only the results based on the mentioned timeframe.}
}
\value{
A data frame consisting of data related to the term belonging to a source 
with in the given time frame is returned.
}
\description{
Get Strength, Sentiment, Passion and Reach of a term in Blogs, MicroBlogs, 
Bookmarks, Images or all of the sources.
}
\details{
\code{term} is any character string
\code{source} can be Blogs, MicroBlogs, Bookmarks, Images, all or list of all
 of the mentioned sources.
\code{time.frame} can be h, 12h, 24h, w, m or all. Meaning in respective order
Last Hour, Last 12 Hours, Last Day, Last Week, Last Month or Anytime.
}
\examples{
socialMention("United_States")

source <- "blogs"
time.frame <- "12h"
socialMention("India", source, time.frame)

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\references{

}
\seealso{
\code{\link{}}
}
\keyword{Mentions}
\keyword{Social}

